+++ core/commons/networks/service_network_config.go
@@ -4,7 +4,16 @@ import (
 	"github.com/kurtosis-tech/kurtosis/commons/docker"
 	"github.com/kurtosis-tech/kurtosis/commons/services"
 	"github.com/palantir/stacktrace"
+	"github.com/sirupsen/logrus"
 )
+
+type serviceConfig struct {
+	// The Docker image that will be used to run the given service
+	// Nil has a special meaning of "use the image being tested"
+	dockerImage *string
+	availabilityChecker services.ServiceAvailabilityChecker
+	initializer services.ServiceInitializer
+}
 
 // Builder to ease the declaration of the network state we want
 type ServiceNetworkConfigBuilder struct {
@@ -26,19 +35,20 @@ type ServiceNetworkConfigBuilder struct {
 	nextServiceId int
 
 	// Factories that will be used to construct the nodes at build time
-	configurations map[int]services.ServiceFactory
+	configurations map[int]serviceConfig
 
-	// Tracks the next service configuration ID that will be doled out upon a call to AddServiceConfiguration
+	// Tracks the next service configuration ID that will be doled out upon a call to AddStaticImageConfiguration
 	nextConfigurationId int
 
 }
 
 func NewServiceNetworkConfigBuilder() *ServiceNetworkConfigBuilder {
+	// TODO use aliased struct types to make it clear which IDs are which
 	serviceConfigs := make(map[int]int)
 	serviceDependencies := make(map[int]map[int]bool)
 	serviceStartOrder := make([]int, 0)
 	onlyDependentServices := make(map[int]bool)
-	configurations := make(map[int]services.ServiceFactory)
+	configurations := make(map[int]serviceConfig)
 	return &ServiceNetworkConfigBuilder{
 		serviceConfigs:      serviceConfigs,
 		serviceDependencies: serviceDependencies,
@@ -50,12 +60,29 @@ func NewServiceNetworkConfigBuilder() *ServiceNetworkConfigBuilder {
 	}
 }
 
-// Adds a service configuration to the network, that can be referenced later with AddService
-func (builder *ServiceNetworkConfigBuilder) AddServiceConfiguration(factory services.ServiceFactory) int {
+// Adds a service configuration to the network that will run a static Docker image
+// This configuration can be referenced later with AddService
+func (builder *ServiceNetworkConfigBuilder) AddStaticImageConfiguration(
+		dockerImage *string,
+		initializerCore services.ServiceInitializerCore,
+		availabilityCheckerCore services.ServiceAvailabilityCheckerCore) int {
+	serviceConfig := serviceConfig{
+		dockerImage: dockerImage,
+		availabilityChecker: *services.NewServiceAvailabilityChecker(availabilityCheckerCore),
+		initializer:         *services.NewServiceInitializer(initializerCore),
+	}
 	configurationId := builder.nextConfigurationId
 	builder.nextConfigurationId = builder.nextConfigurationId + 1
-	builder.configurations[configurationId] = factory
+	builder.configurations[configurationId] = serviceConfig
 	return configurationId
+}
+
+// Adds a service configuration to the network that will run the Docker image being tested
+// This configuration can be referenced later with AddService
+func (builder *ServiceNetworkConfigBuilder) AddTestImageConfiguration(
+		initializerCore services.ServiceInitializerCore,
+		availabilityCheckerCore services.ServiceAvailabilityCheckerCore) int {
+	return builder.AddStaticImageConfiguration(nil, initializerCore, availabilityCheckerCore)
 }
 
 // Adds a serivce to the graph, with the specified dependencies (with the map used only as a set - the values are ignored)
@@ -121,9 +148,9 @@ func (builder ServiceNetworkConfigBuilder) Build() *ServiceNetworkConfig {
 		onlyDependentServicesCopy[dependencyId] = true
 	}
 
-	configurationsCopy := make(map[int]services.ServiceFactory)
-	for configurationId, factory := range builder.configurations {
-		configurationsCopy[configurationId] = factory
+	configurationsCopy := make(map[int]serviceConfig)
+	for configurationId, config := range builder.configurations {
+		configurationsCopy[configurationId] = config
 	}
 
 	return &ServiceNetworkConfig{
@@ -144,7 +171,7 @@ type ServiceNetworkConfig struct {
 	// push that to the implementer of the network (make them do the calls based off what they know)
 	// Don't want to rip it out yet though because it was a pain to put in
 	onlyDependentServices map[int]bool
-	configurations map[int]services.ServiceFactory
+	configurations map[int]serviceConfig
 }
 
 /*
@@ -156,21 +183,21 @@ Returns:
 		user of this method should check if the error result is set and, if so, shut down the running containers!
  */
 // TODO use the network name to create a new network!!
-func (networkCfg ServiceNetworkConfig) CreateAndRun(publicIpProvider *FreeIpAddrTracker, manager *docker.DockerManager) (*RawServiceNetwork, error) {
+func (networkCfg ServiceNetworkConfig) CreateNetwork(testImage string, publicIpProvider *FreeIpAddrTracker, manager *docker.DockerManager) (*RawServiceNetwork, error) {
 	runningServices := make(map[int]services.Service)
 	serviceIps := make(map[int]string)
 	serviceContainerIds := make(map[int]string)
+	allServiceDependencies := make(map[int][]services.Service)
 
+	// First pass: start all services
+	logrus.Info("Creating & starting test network containers...")
 	for _, serviceId := range networkCfg.servicesStartOrder {
-		serviceDependenciesIds := networkCfg.serviceDependencies[serviceId]
-		serviceDependencies := make([]services.Service, 0, len(serviceDependenciesIds))
-		for dependencyId, _ := range serviceDependenciesIds {
-			// We're guaranteed that this dependency will already be running due to the ordering we enforce in the builder
-			serviceDependencies = append(serviceDependencies, runningServices[dependencyId])
-		}
+		serviceDependencies := networkCfg.getServiceDependencies(serviceId, runningServices)
 
-		configId := networkCfg.serviceConfigs[serviceId]
-		factory := networkCfg.configurations[configId]
+		config, err := networkCfg.getServiceConfig(serviceId)
+		if err != nil {
+			return nil, stacktrace.Propagate(err, "Could not get service config for service ID %v", serviceId)
+		}
 
 		staticIp, err := publicIpProvider.GetFreeIpAddr()
 		if err != nil {
@@ -180,21 +207,103 @@ func (networkCfg ServiceNetworkConfig) CreateAndRun(publicIpProvider *FreeIpAddr
 			}, stacktrace.Propagate(err, "Failed to allocate static IP for service %d", serviceId)
 		}
 
-		service, containerId, err := factory.Construct(staticIp, manager, serviceDependencies)
+		dockerImagePtr := config.dockerImage
+		if dockerImagePtr == nil {
+			dockerImagePtr = &testImage
+		}
+
+		service, containerId, err := config.initializer.CreateService(*dockerImagePtr, staticIp, manager, serviceDependencies)
 		if err != nil {
 			return &RawServiceNetwork{
 				ServiceIPs:   serviceIps,
 				ContainerIds: serviceContainerIds,
-			}, stacktrace.Propagate(err, "Failed to construct service from factory")
+			}, stacktrace.Propagate(err, "Failed to construct service from serviceConfig")
 		}
 
 		runningServices[serviceId] = service
 		serviceIps[serviceId] = staticIp
 		serviceContainerIds[serviceId] = containerId
+		allServiceDependencies[serviceId] = serviceDependencies
 	}
+	logrus.Info("Test network containers created & started")
 
 	return &RawServiceNetwork{
-		ContainerIds:   serviceContainerIds,
-		ServiceIPs: serviceIps,
+		ServiceIPs:   serviceIps,
+		ContainerIds: serviceContainerIds,
 	}, nil
+}
+
+// Intended for use by the test controller
+// Reads basic information about the nodes in the network and uses the topology information stored
+//  in this object to:
+//		1) translate each IP into the appropriate service using the constructors given in this object and
+//		2) wait until all services are available
+func (networkCfg ServiceNetworkConfig) LoadNetwork(rawInfo RawServiceNetwork) (map[int]services.Service, error) {
+	// First pass: construct the instantions of each service object
+	logrus.Info("Loading services from IPs...")
+	runningServices := make(map[int]services.Service)
+	for _, serviceId := range networkCfg.servicesStartOrder {
+
+		ipAddr, found := rawInfo.ServiceIPs[serviceId]
+		if !found {
+			return nil, stacktrace.NewError("Missing expected service ID '%v' in network info", serviceId)
+		}
+
+		config, err := networkCfg.getServiceConfig(serviceId)
+		if err != nil {
+			return nil, stacktrace.Propagate(err, "Could not get service config for service ID %v", serviceId)
+		}
+
+		service := config.initializer.LoadService(ipAddr)
+		runningServices[serviceId] = service
+	}
+	logrus.Info("All services loaded from IPs")
+
+	// Second pass: wait for all services to come up
+	logrus.Info("Waiting for network to become available...")
+	for _, serviceId := range networkCfg.servicesStartOrder {
+		service := runningServices[serviceId]
+		serviceDependencies := networkCfg.getServiceDependencies(serviceId, runningServices)
+		config, err := networkCfg.getServiceConfig(serviceId)
+		if err != nil {
+			return nil, stacktrace.Propagate(err, "Could not get service config for ID %v", serviceId)
+		}
+
+		logrus.Debugf("Waiting for service %v to become available...", serviceId)
+		if err := config.availabilityChecker.WaitForStartup(service, serviceDependencies); err != nil {
+			return nil, stacktrace.Propagate(err, "An error occurred waiting for service %v to start up", serviceId)
+		}
+		logrus.Debugf("Service %v is available", serviceId)
+	}
+	logrus.Info("Network is available")
+
+	return runningServices, nil
+}
+
+// Convenience function to get a service's dependencies as a []Service
+func (networkCfg ServiceNetworkConfig) getServiceDependencies(serviceId int, runningServices map[int]services.Service) []services.Service {
+	serviceDependenciesIds := networkCfg.serviceDependencies[serviceId]
+	serviceDependencies := make([]services.Service, 0, len(serviceDependenciesIds))
+	for dependencyId, _ := range serviceDependenciesIds {
+		// We're guaranteed that this dependency will already be running due to the ordering we enforce in the builder
+		serviceDependencies = append(serviceDependencies, runningServices[dependencyId])
+	}
+	return serviceDependencies
+}
+
+// Convenience function to get a service's serviceConfig
+func (networkCfg ServiceNetworkConfig) getServiceConfig(serviceId int) (serviceConfig, error) {
+	configId, found := networkCfg.serviceConfigs[serviceId]
+	if !found {
+		return serviceConfig{}, stacktrace.NewError("Found ID '%v' in the network info but no configuration is defined for this ID in the network config", serviceId)
+	}
+
+	config, found := networkCfg.configurations[configId]
+	if !found {
+		return serviceConfig{}, stacktrace.NewError(
+			"Service ID '%v' uses service configuration '%v', but this service config wasn't found in this network configuration; this is likely a code problem",
+			serviceId,
+			configId)
+	}
+	return config, nil
 }