+++ core/commons/services/service_initializer_core.go
@@ -2,18 +2,18 @@ package services
 
 import "os"
 
-// Contains configuration determining what type of objects the ServiceFactory will produce
+// TODO When Go has generics, parameterize this to be <N, S extends N> where S is the
+//  specific service interface and N represents the interface that every node on the network has
+// Contains configuration determining what type of objects the ServiceInitializer will produce
 // This is implicitly a DockerContainerServiceFactoryConfig; we could abstract it easily if we wanted other foundations for services
-type ServiceFactoryConfig interface {
-	GetDockerImage() string
-
+type ServiceInitializerCore interface {
 	GetUsedPorts() map[int]bool
 
-	// TODO when Go gets generics, make the type of 'dependencies' be the same as the output of GetStartCommand
+	// TODO when Go gets generics, make the type of 'dependencies' to be []N
 	// If Go had generics, dependencies should be of type []T
 	GetStartCommand(publicIpAddr string, dependencies []Service) []string
 
-	// If Go had generics, the return type would be T
+	// TODO When Go has generics, make this return type to be S
 	GetServiceFromIp(ipAddr string) Service
 
 	GetFilepathsToMount() map[string]bool