+++ core/CHANGELOG.md
@@ -1,2 +1,7 @@
 # TBD
 * Stop the node network after the test controller runs
+* Rename ServiceFactory(,Config) -> ServiceInitializer(,Core)
+* Fix bug with not actually catching panics when running a test
+* Fix a bug with the TestController not appropriately catching panics
+* Log test result as soon as the test is finished
+* Add some extra unit tests